+++ python/tvm/dlight/gpu/__init__.py
@@ -18,8 +18,9 @@
 GPU-generic schedule rules.
 For CUDA/ROCm/Vulkan/Metal-specific rules, use `tvm.dlight.cuda/rocm/vulkan/metal` instead
 """
-from .decode_gemv import DecodeGEMV
+from .gemv import GEMV
 from .fallback import Fallback
 from .matmul import Matmul
 from .reduction import Reduction
 from .transpose import Transpose
+from .general_reduction import GeneralReduction
